change Ducks Pattern
