+++ redux-start/src/redux/store.js
@@ -1,5 +1,5 @@
 import { applyMiddleware, createStore } from "redux";
-import todoApp from "./reducers/reducer";
+import todoApp from "./modules/reducer";
 import {composeWithDevTools} from 'redux-devtools-extension'
 import thunk from 'redux-thunk'
 import promise from "redux-promise-middleware"
